Refactoring utilities functions to a separate folder.
