+++ pkg/azuredisk/azure_common_test.go
@@ -23,11 +23,6 @@ import (
 	"strings"
 	"testing"
 	"time"
-
-	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2020-12-01/compute"
-	"github.com/stretchr/testify/assert"
-
-	v1 "k8s.io/api/core/v1"
 )
 
 type fakeFileInfo struct {
@@ -134,224 +129,6 @@ func TestIoHandler(t *testing.T) {
 		if disk != "/dev/"+devName || err != nil {
 			t.Errorf("no data disk found: disk %v err %v", disk, err)
 		}
-	}
-}
-
-func TestNormalizeStorageAccountType(t *testing.T) {
-	tests := []struct {
-		cloud                  string
-		storageAccountType     string
-		disableAzureStackCloud bool
-		expectedAccountType    compute.DiskStorageAccountTypes
-		expectError            bool
-	}{
-		{
-			cloud:                  azurePublicCloud,
-			storageAccountType:     "",
-			disableAzureStackCloud: false,
-			expectedAccountType:    compute.StandardSSDLRS,
-			expectError:            false,
-		},
-		{
-			cloud:                  azureStackCloud,
-			storageAccountType:     "",
-			disableAzureStackCloud: false,
-			expectedAccountType:    compute.StandardLRS,
-			expectError:            false,
-		},
-		{
-			cloud:                  azurePublicCloud,
-			storageAccountType:     "NOT_EXISTING",
-			disableAzureStackCloud: false,
-			expectedAccountType:    "",
-			expectError:            true,
-		},
-		{
-			cloud:                  azurePublicCloud,
-			storageAccountType:     "Standard_LRS",
-			disableAzureStackCloud: false,
-			expectedAccountType:    compute.StandardLRS,
-			expectError:            false,
-		},
-		{
-			cloud:                  azurePublicCloud,
-			storageAccountType:     "Premium_LRS",
-			disableAzureStackCloud: false,
-			expectedAccountType:    compute.PremiumLRS,
-			expectError:            false,
-		},
-		{
-			cloud:                  azurePublicCloud,
-			storageAccountType:     "StandardSSD_LRS",
-			disableAzureStackCloud: false,
-			expectedAccountType:    compute.StandardSSDLRS,
-			expectError:            false,
-		},
-		{
-			cloud:                  azurePublicCloud,
-			storageAccountType:     "UltraSSD_LRS",
-			disableAzureStackCloud: false,
-			expectedAccountType:    compute.UltraSSDLRS,
-			expectError:            false,
-		},
-		{
-			cloud:                  azureStackCloud,
-			storageAccountType:     "UltraSSD_LRS",
-			disableAzureStackCloud: false,
-			expectedAccountType:    "",
-			expectError:            true,
-		},
-		{
-			cloud:                  azureStackCloud,
-			storageAccountType:     "UltraSSD_LRS",
-			disableAzureStackCloud: true,
-			expectedAccountType:    compute.UltraSSDLRS,
-			expectError:            false,
-		},
-	}
-
-	for _, test := range tests {
-		result, err := normalizeStorageAccountType(test.storageAccountType, test.cloud, test.disableAzureStackCloud)
-		assert.Equal(t, result, test.expectedAccountType)
-		assert.Equal(t, err != nil, test.expectError, fmt.Sprintf("error msg: %v", err))
-	}
-}
-
-func TestNormalizeCachingMode(t *testing.T) {
-	tests := []struct {
-		desc          string
-		req           v1.AzureDataDiskCachingMode
-		expectedErr   error
-		expectedValue v1.AzureDataDiskCachingMode
-	}{
-		{
-			desc:          "CachingMode not exist",
-			req:           "",
-			expectedErr:   nil,
-			expectedValue: v1.AzureDataDiskCachingReadOnly,
-		},
-		{
-			desc:          "Not supported CachingMode",
-			req:           "WriteOnly",
-			expectedErr:   fmt.Errorf("azureDisk - WriteOnly is not supported cachingmode. Supported values are [None ReadOnly ReadWrite]"),
-			expectedValue: "",
-		},
-		{
-			desc:          "Valid CachingMode",
-			req:           "ReadOnly",
-			expectedErr:   nil,
-			expectedValue: "ReadOnly",
-		},
-	}
-	for _, test := range tests {
-		value, err := normalizeCachingMode(test.req)
-		assert.Equal(t, value, test.expectedValue)
-		assert.Equal(t, err, test.expectedErr, fmt.Sprintf("error msg: %v", err))
-	}
-}
-
-func TestNormalizeNetworkAccessPolicy(t *testing.T) {
-	tests := []struct {
-		networkAccessPolicy         string
-		expectedNetworkAccessPolicy compute.NetworkAccessPolicy
-		expectError                 bool
-	}{
-		{
-			networkAccessPolicy:         "",
-			expectedNetworkAccessPolicy: compute.AllowAll,
-			expectError:                 false,
-		},
-		{
-			networkAccessPolicy:         "AllowAll",
-			expectedNetworkAccessPolicy: compute.AllowAll,
-			expectError:                 false,
-		},
-		{
-			networkAccessPolicy:         "DenyAll",
-			expectedNetworkAccessPolicy: compute.DenyAll,
-			expectError:                 false,
-		},
-		{
-			networkAccessPolicy:         "AllowPrivate",
-			expectedNetworkAccessPolicy: compute.AllowPrivate,
-			expectError:                 false,
-		},
-		{
-			networkAccessPolicy:         "allowAll",
-			expectedNetworkAccessPolicy: compute.NetworkAccessPolicy(""),
-			expectError:                 true,
-		},
-		{
-			networkAccessPolicy:         "invalid",
-			expectedNetworkAccessPolicy: compute.NetworkAccessPolicy(""),
-			expectError:                 true,
-		},
-	}
-
-	for _, test := range tests {
-		result, err := normalizeNetworkAccessPolicy(test.networkAccessPolicy)
-		assert.Equal(t, result, test.expectedNetworkAccessPolicy)
-		assert.Equal(t, err != nil, test.expectError, fmt.Sprintf("error msg: %v", err))
-	}
-}
-
-func TestGetDiskLUN(t *testing.T) {
-	tests := []struct {
-		deviceInfo  string
-		expectedLUN int32
-		expectError bool
-	}{
-		{
-			deviceInfo:  "0",
-			expectedLUN: 0,
-			expectError: false,
-		},
-		{
-			deviceInfo:  "10",
-			expectedLUN: 10,
-			expectError: false,
-		},
-		{
-			deviceInfo:  "11d",
-			expectedLUN: -1,
-			expectError: true,
-		},
-		{
-			deviceInfo:  "999",
-			expectedLUN: -1,
-			expectError: true,
-		},
-		{
-			deviceInfo:  "",
-			expectedLUN: -1,
-			expectError: true,
-		},
-		{
-			deviceInfo:  "/dev/disk/azure/scsi1/lun2",
-			expectedLUN: 2,
-			expectError: false,
-		},
-		{
-			deviceInfo:  "/dev/disk/azure/scsi0/lun12",
-			expectedLUN: 12,
-			expectError: false,
-		},
-		{
-			deviceInfo:  "/devhost/disk/azure/scsi0/lun13",
-			expectedLUN: 13,
-			expectError: false,
-		},
-		{
-			deviceInfo:  "/dev/disk/by-id/scsi1/lun2",
-			expectedLUN: -1,
-			expectError: true,
-		},
-	}
-
-	for _, test := range tests {
-		result, err := getDiskLUN(test.deviceInfo)
-		assert.Equal(t, result, test.expectedLUN)
-		assert.Equal(t, err != nil, test.expectError, fmt.Sprintf("error msg: %v", err))
 	}
 }
