+++ pkg/azuredisk/azuredisk_test.go
@@ -18,585 +18,18 @@ package azuredisk
 
 import (
 	"context"
-	"fmt"
 	"io/ioutil"
 	"os"
-	"path/filepath"
-	"reflect"
 	"testing"
 
 	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2020-12-01/compute"
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/grpc/status"
+	constants "sigs.k8s.io/azuredisk-csi-driver/pkg/azureutils"
 	"sigs.k8s.io/cloud-provider-azure/pkg/azureclients/diskclient/mockdiskclient"
 	azure "sigs.k8s.io/cloud-provider-azure/pkg/provider"
 )
-
-func TestGetDiskName(t *testing.T) {
-	mDiskPathRE := managedDiskPathRE
-	tests := []struct {
-		options   string
-		expected1 string
-		expected2 error
-	}{
-		{
-			options:   "testurl/subscriptions/12/resourceGroups/23/providers/Microsoft.Compute/disks/name",
-			expected1: "name",
-			expected2: nil,
-		},
-		{
-			options:   "testurl/subscriptions/23/providers/Microsoft.Compute/disks/name",
-			expected1: "",
-			expected2: fmt.Errorf("could not get disk name from testurl/subscriptions/23/providers/Microsoft.Compute/disks/name, correct format: %s", mDiskPathRE),
-		},
-		{
-			options:   "testurl/subscriptions/12/resourcegroups/23/providers/microsoft.compute/disks/name",
-			expected1: "name",
-			expected2: nil,
-		},
-		{
-			options:   "testurl/subscriPtions/12/Resourcegroups/23/Providers/microsoft.compute/dISKS/name",
-			expected1: "name",
-			expected2: nil,
-		},
-		{
-			options:   "http://test.com/vhds/name",
-			expected1: "",
-			expected2: fmt.Errorf("could not get disk name from http://test.com/vhds/name, correct format: %s", mDiskPathRE),
-		},
-		{
-			options:   "http://test.io/name",
-			expected1: "",
-			expected2: fmt.Errorf("could not get disk name from http://test.io/name, correct format: %s", mDiskPathRE),
-		},
-	}
-
-	for _, test := range tests {
-		result1, result2 := GetDiskName(test.options)
-		if !reflect.DeepEqual(result1, test.expected1) || !reflect.DeepEqual(result2, test.expected2) {
-			t.Errorf("input: %q, getDiskName result1: %q, expected1: %q, result2: %q, expected2: %q", test.options, result1, test.expected1,
-				result2, test.expected2)
-		}
-	}
-}
-
-func TestGetSnapshotName(t *testing.T) {
-	tests := []struct {
-		options   string
-		expected1 string
-		expected2 error
-	}{
-		{
-			options:   "testurl/subscriptions/12/resourceGroups/23/providers/Microsoft.Compute/snapshots/snapshot-name",
-			expected1: "snapshot-name",
-			expected2: nil,
-		},
-		{
-			options:   "testurl/subscriptions/12/resourcegroups/23/providers/microsoft.compute/SNAPSHOTS/snapshot-name",
-			expected1: "snapshot-name",
-			expected2: nil,
-		},
-		{
-			options:   "testurl/subscriptions/23/providers/Microsoft.Compute/snapshots/snapshot-name",
-			expected1: "",
-			expected2: fmt.Errorf("could not get snapshot name from testurl/subscriptions/23/providers/Microsoft.Compute/snapshots/snapshot-name, correct format: %s", diskSnapshotPathRE),
-		},
-	}
-
-	for _, test := range tests {
-		result1, result2 := getSnapshotName(test.options)
-		if !reflect.DeepEqual(result1, test.expected1) || !reflect.DeepEqual(result2, test.expected2) {
-			t.Errorf("input: %q, getSnapshotName result1: %q, expected1: %q, result2: %q, expected2: %q", test.options, result1, test.expected1,
-				result2, test.expected2)
-		}
-	}
-}
-
-func TestGetResourceGroupFromURI(t *testing.T) {
-	tests := []struct {
-		diskURL        string
-		expectedResult string
-		expectError    bool
-	}{
-		{
-			diskURL:        "/subscriptions/4be8920b-2978-43d7-axyz-04d8549c1d05/resourceGroups/azure-k8s1102/providers/Microsoft.Compute/disks/andy-mghyb1102-dynamic-pvc-f7f014c9-49f4-11e8-ab5c-000d3af7b38e",
-			expectedResult: "azure-k8s1102",
-			expectError:    false,
-		},
-		{
-			// case insentive check
-			diskURL:        "/subscriptions/4be8920b-2978-43d7-axyz-04d8549c1d05/resourcegroups/azure-k8s1102/providers/Microsoft.Compute/disks/andy-mghyb1102-dynamic-pvc-f7f014c9-49f4-11e8-ab5c-000d3af7b38e",
-			expectedResult: "azure-k8s1102",
-			expectError:    false,
-		},
-		{
-			diskURL:        "/4be8920b-2978-43d7-axyz-04d8549c1d05/resourceGroups/azure-k8s1102/providers/Microsoft.Compute/disks/andy-mghyb1102-dynamic-pvc-f7f014c9-49f4-11e8-ab5c-000d3af7b38e",
-			expectedResult: "",
-			expectError:    true,
-		},
-		{
-			diskURL:        "",
-			expectedResult: "",
-			expectError:    true,
-		},
-	}
-
-	for _, test := range tests {
-		result, err := GetResourceGroupFromURI(test.diskURL)
-		assert.Equal(t, result, test.expectedResult, "Expect result not equal with getResourceGroupFromURI(%s) return: %q, expected: %q",
-			test.diskURL, result, test.expectedResult)
-
-		if test.expectError {
-			assert.NotNil(t, err, "Expect error during getResourceGroupFromURI(%s)", test.diskURL)
-		} else {
-			assert.Nil(t, err, "Expect error is nil during getResourceGroupFromURI(%s)", test.diskURL)
-		}
-	}
-}
-
-func TestIsValidDiskURI(t *testing.T) {
-	supportedManagedDiskURI := diskURISupportedManaged
-
-	tests := []struct {
-		diskURI     string
-		expectError error
-	}{
-		{
-			diskURI:     "/subscriptions/b9d2281e/resourceGroups/test-resource/providers/Microsoft.Compute/disks/pvc-disk-dynamic-9e102c53",
-			expectError: nil,
-		},
-		{
-			diskURI:     "/Subscriptions/b9d2281e/resourceGroups/test-resource/providers/Microsoft.Compute/disks/pvc-disk-dynamic-9e102c53",
-			expectError: nil,
-		},
-		{
-			diskURI:     "resourceGroups/test-resource/providers/Microsoft.Compute/disks/pvc-disk-dynamic-9e102c53",
-			expectError: fmt.Errorf("Inavlid DiskURI: resourceGroups/test-resource/providers/Microsoft.Compute/disks/pvc-disk-dynamic-9e102c53, correct format: %v", supportedManagedDiskURI),
-		},
-		{
-			diskURI:     "https://test-saccount.blob.core.windows.net/container/pvc-disk-dynamic-9e102c53-593d-11e9-934e-705a0f18a318.vhd",
-			expectError: fmt.Errorf("Inavlid DiskURI: https://test-saccount.blob.core.windows.net/container/pvc-disk-dynamic-9e102c53-593d-11e9-934e-705a0f18a318.vhd, correct format: %v", supportedManagedDiskURI),
-		},
-		{
-			diskURI:     "test.com",
-			expectError: fmt.Errorf("Inavlid DiskURI: test.com, correct format: %v", supportedManagedDiskURI),
-		},
-		{
-			diskURI:     "http://test-saccount.blob.core.windows.net/container/pvc-disk-dynamic-9e102c53-593d-11e9-934e-705a0f18a318.vhd",
-			expectError: fmt.Errorf("Inavlid DiskURI: http://test-saccount.blob.core.windows.net/container/pvc-disk-dynamic-9e102c53-593d-11e9-934e-705a0f18a318.vhd, correct format: %v", supportedManagedDiskURI),
-		},
-	}
-
-	for _, test := range tests {
-		err := isValidDiskURI(test.diskURI)
-		if !reflect.DeepEqual(err, test.expectError) {
-			t.Errorf("DiskURI: %q, isValidDiskURI err: %q, expected1: %q", test.diskURI, err, test.expectError)
-		}
-	}
-}
-
-func TestIsARMResourceID(t *testing.T) {
-	tests := []struct {
-		resourceID   string
-		expectResult bool
-	}{
-		{
-			resourceID:   "/subscriptions/b9d2281e/resourceGroups/test-resource/providers/Microsoft.Compute/disks/pvc-disk-dynamic-9e102c53",
-			expectResult: true,
-		},
-		{
-			resourceID:   "/Subscriptions/b9d2281e/resourceGroups/test-resource/providers/Microsoft.Compute/disks/pvc-disk-dynamic-9e102c53",
-			expectResult: true,
-		},
-		{
-			resourceID:   "resourceGroups/test-resource/providers/Microsoft.Compute/disks/pvc-disk-dynamic-9e102c53",
-			expectResult: false,
-		},
-		{
-			resourceID:   "https://test-saccount.blob.core.windows.net/container/pvc-disk-dynamic-9e102c53-593d-11e9-934e-705a0f18a318.vhd",
-			expectResult: false,
-		},
-		{
-			resourceID:   "test.com",
-			expectResult: false,
-		},
-		{
-			resourceID:   "",
-			expectResult: false,
-		},
-	}
-
-	for _, test := range tests {
-		result := isARMResourceID(test.resourceID)
-		if result != test.expectResult {
-			t.Errorf("ResourceID: %s, result: %v, expectResult: %v", test.resourceID, result, test.expectResult)
-		}
-	}
-}
-
-func TestGetValidDiskName(t *testing.T) {
-	tests := []struct {
-		volumeName string
-		expected   string
-	}{
-		{
-			volumeName: "az",
-			expected:   "az",
-		},
-		{
-			volumeName: "09",
-			expected:   "09",
-		},
-		{
-			volumeName: "a-z",
-			expected:   "a-z",
-		},
-		{
-			volumeName: "AZ",
-			expected:   "AZ",
-		},
-		{
-			volumeName: "123456789-123456789-123456789-123456789-123456789.123456789-123456789_1234567890",
-			expected:   "123456789-123456789-123456789-123456789-123456789.123456789-123456789_1234567890",
-		},
-		{
-			volumeName: "123456789-123456789-123456789-123456789-123456789.123456789-123456789_1234567890-123456789-123456789-123456789-123456789-123456789.123456789-123456789_1234567890-123456789-123456789-123456789-123456789-123456789.123456789-123456789_1234567890",
-			expected:   "123456789-123456789-123456789-123456789-123456789.123456789-123456789_1234567890",
-		},
-	}
-
-	for _, test := range tests {
-		result := getValidDiskName(test.volumeName)
-		if !reflect.DeepEqual(result, test.expected) {
-			t.Errorf("input: %q, getValidFileShareName result: %q, expected: %q", test.volumeName, result, test.expected)
-		}
-	}
-}
-
-func TestCheckDiskName(t *testing.T) {
-	tests := []struct {
-		diskName string
-		expected bool
-	}{
-		{
-			diskName: "a",
-			expected: true,
-		},
-		{
-			diskName: ".",
-			expected: false,
-		},
-		{
-			diskName: "_",
-			expected: false,
-		},
-		{
-			diskName: "_",
-			expected: false,
-		},
-		{
-			diskName: "09",
-			expected: true,
-		},
-		{
-			diskName: "az",
-			expected: true,
-		},
-		{
-			diskName: "1_",
-			expected: true,
-		},
-		{
-			diskName: "_1",
-			expected: false,
-		},
-		{
-			diskName: "1.",
-			expected: false,
-		},
-		{
-			diskName: "1-",
-			expected: false,
-		},
-		{
-			diskName: "0.z",
-			expected: true,
-		},
-		{
-			diskName: "1.2",
-			expected: true,
-		},
-		{
-			diskName: "a-9",
-			expected: true,
-		},
-		{
-			diskName: "a_c",
-			expected: true,
-		},
-		{
-			diskName: "1__",
-			expected: true,
-		},
-		{
-			diskName: "a---9",
-			expected: true,
-		},
-		{
-			diskName: "1#2",
-			expected: false,
-		},
-		{
-			diskName: "-",
-			expected: false,
-		},
-		{
-			diskName: "test",
-			expected: true,
-		},
-	}
-
-	for _, test := range tests {
-		result := checkDiskName(test.diskName)
-		if !reflect.DeepEqual(result, test.expected) {
-			t.Errorf("input: %q, checkShareNameBeginAndEnd result: %v, expected: %v", test.diskName, result, test.expected)
-		}
-	}
-}
-
-func TestGetSourceVolumeID(t *testing.T) {
-	SourceResourceID := "test"
-
-	tests := []struct {
-		snapshot *compute.Snapshot
-		expected string
-	}{
-		{
-			snapshot: &compute.Snapshot{
-				SnapshotProperties: &compute.SnapshotProperties{
-					CreationData: &compute.CreationData{
-						SourceResourceID: &SourceResourceID,
-					},
-				},
-			},
-			expected: "test",
-		},
-		{
-			snapshot: &compute.Snapshot{
-				SnapshotProperties: &compute.SnapshotProperties{
-					CreationData: &compute.CreationData{},
-				},
-			},
-			expected: "",
-		},
-		{
-			snapshot: &compute.Snapshot{
-				SnapshotProperties: &compute.SnapshotProperties{},
-			},
-			expected: "",
-		},
-		{
-			snapshot: &compute.Snapshot{},
-			expected: "",
-		},
-		{
-			snapshot: nil,
-			expected: "",
-		},
-	}
-
-	for _, test := range tests {
-		result := getSourceVolumeID(test.snapshot)
-		if !reflect.DeepEqual(result, test.expected) {
-			t.Errorf("input: %v, getValidFileShareName result: %q, expected: %q", test.snapshot, result, test.expected)
-		}
-	}
-}
-
-func TestGetValidCreationData(t *testing.T) {
-	sourceResourceSnapshotID := "/subscriptions/xxx/resourceGroups/xxx/providers/Microsoft.Compute/snapshots/xxx"
-	sourceResourceVolumeID := "/subscriptions/xxx/resourceGroups/xxx/providers/Microsoft.Compute/disks/xxx"
-
-	tests := []struct {
-		subscriptionID   string
-		resourceGroup    string
-		sourceResourceID string
-		sourceType       string
-		expected1        compute.CreationData
-		expected2        error
-	}{
-		{
-			subscriptionID:   "",
-			resourceGroup:    "",
-			sourceResourceID: "",
-			sourceType:       "",
-			expected1: compute.CreationData{
-				CreateOption: compute.Empty,
-			},
-			expected2: nil,
-		},
-		{
-			subscriptionID:   "",
-			resourceGroup:    "",
-			sourceResourceID: "/subscriptions/xxx/resourceGroups/xxx/providers/Microsoft.Compute/snapshots/xxx",
-			sourceType:       sourceSnapshot,
-			expected1: compute.CreationData{
-				CreateOption:     compute.Copy,
-				SourceResourceID: &sourceResourceSnapshotID,
-			},
-			expected2: nil,
-		},
-		{
-			subscriptionID:   "xxx",
-			resourceGroup:    "xxx",
-			sourceResourceID: "xxx",
-			sourceType:       sourceSnapshot,
-			expected1: compute.CreationData{
-				CreateOption:     compute.Copy,
-				SourceResourceID: &sourceResourceSnapshotID,
-			},
-			expected2: nil,
-		},
-		{
-			subscriptionID:   "",
-			resourceGroup:    "",
-			sourceResourceID: "/subscriptions/23/providers/Microsoft.Compute/disks/name",
-			sourceType:       sourceSnapshot,
-			expected1:        compute.CreationData{},
-			expected2:        fmt.Errorf("sourceResourceID(%s) is invalid, correct format: %s", "/subscriptions//resourceGroups//providers/Microsoft.Compute/snapshots//subscriptions/23/providers/Microsoft.Compute/disks/name", diskSnapshotPathRE),
-		},
-		{
-			subscriptionID:   "",
-			resourceGroup:    "",
-			sourceResourceID: "http://test.com/vhds/name",
-			sourceType:       sourceSnapshot,
-			expected1:        compute.CreationData{},
-			expected2:        fmt.Errorf("sourceResourceID(%s) is invalid, correct format: %s", "/subscriptions//resourceGroups//providers/Microsoft.Compute/snapshots/http://test.com/vhds/name", diskSnapshotPathRE),
-		},
-		{
-			subscriptionID:   "",
-			resourceGroup:    "",
-			sourceResourceID: "/subscriptions/xxx/snapshots/xxx",
-			sourceType:       sourceSnapshot,
-			expected1:        compute.CreationData{},
-			expected2:        fmt.Errorf("sourceResourceID(%s) is invalid, correct format: %s", "/subscriptions//resourceGroups//providers/Microsoft.Compute/snapshots//subscriptions/xxx/snapshots/xxx", diskSnapshotPathRE),
-		},
-		{
-			subscriptionID:   "",
-			resourceGroup:    "",
-			sourceResourceID: "/subscriptions/xxx/resourceGroups/xxx/providers/Microsoft.Compute/snapshots/xxx/snapshots/xxx/snapshots/xxx",
-			sourceType:       sourceSnapshot,
-			expected1:        compute.CreationData{},
-			expected2:        fmt.Errorf("sourceResourceID(%s) is invalid, correct format: %s", "/subscriptions/xxx/resourceGroups/xxx/providers/Microsoft.Compute/snapshots/xxx/snapshots/xxx/snapshots/xxx", diskSnapshotPathRE),
-		},
-		{
-			subscriptionID:   "",
-			resourceGroup:    "",
-			sourceResourceID: "xxx",
-			sourceType:       "",
-			expected1: compute.CreationData{
-				CreateOption: compute.Empty,
-			},
-			expected2: nil,
-		},
-		{
-			subscriptionID:   "",
-			resourceGroup:    "",
-			sourceResourceID: "/subscriptions/xxx/resourceGroups/xxx/providers/Microsoft.Compute/disks/xxx",
-			sourceType:       sourceVolume,
-			expected1: compute.CreationData{
-				CreateOption:     compute.Copy,
-				SourceResourceID: &sourceResourceVolumeID,
-			},
-			expected2: nil,
-		},
-		{
-			subscriptionID:   "xxx",
-			resourceGroup:    "xxx",
-			sourceResourceID: "xxx",
-			sourceType:       sourceVolume,
-			expected1: compute.CreationData{
-				CreateOption:     compute.Copy,
-				SourceResourceID: &sourceResourceVolumeID,
-			},
-			expected2: nil,
-		},
-		{
-			subscriptionID:   "",
-			resourceGroup:    "",
-			sourceResourceID: "/subscriptions/xxx/resourceGroups/xxx/providers/Microsoft.Compute/snapshots/xxx",
-			sourceType:       sourceVolume,
-			expected1:        compute.CreationData{},
-			expected2:        fmt.Errorf("sourceResourceID(%s) is invalid, correct format: %s", "/subscriptions//resourceGroups//providers/Microsoft.Compute/disks//subscriptions/xxx/resourceGroups/xxx/providers/Microsoft.Compute/snapshots/xxx", managedDiskPathRE),
-		},
-	}
-
-	for _, test := range tests {
-		result, err := getValidCreationData(test.subscriptionID, test.resourceGroup, test.sourceResourceID, test.sourceType)
-		if !reflect.DeepEqual(result, test.expected1) || !reflect.DeepEqual(err, test.expected2) {
-			t.Errorf("input sourceResourceID: %v, sourceType: %v, getValidCreationData result: %v, expected1 : %v, err: %v, expected2: %v", test.sourceResourceID, test.sourceType, result, test.expected1, err, test.expected2)
-		}
-	}
-}
-
-func TestIsAvailabilityZone(t *testing.T) {
-	region := "eastus"
-	tests := []struct {
-		desc     string
-		zone     string
-		expected bool
-	}{
-		{"empty string should return false", "", false},
-		{"wrong farmat should return false", "123", false},
-		{"wrong location should return false", "chinanorth-1", false},
-		{"correct zone should return true", "eastus-1", true},
-	}
-
-	for _, test := range tests {
-		actual := isAvailabilityZone(test.zone, region)
-		if actual != test.expected {
-			t.Errorf("test [%q] get unexpected result: %v != %v", test.desc, actual, test.expected)
-		}
-	}
-}
-
-func TestIsCorruptedDir(t *testing.T) {
-	// skip due to permission issues
-	skipIfTestingOnWindows(t)
-	existingMountPath, err := ioutil.TempDir(os.TempDir(), "csi-mount-test")
-	if err != nil {
-		t.Fatalf("failed to create tmp dir: %v", err)
-	}
-	defer os.RemoveAll(existingMountPath)
-
-	curruptedPath := filepath.Join(existingMountPath, "curruptedPath")
-	if err := os.Symlink(existingMountPath, curruptedPath); err != nil {
-		t.Fatalf("failed to create curruptedPath: %v", err)
-	}
-
-	tests := []struct {
-		desc           string
-		dir            string
-		expectedResult bool
-	}{
-		{
-			desc:           "NotExist dir",
-			dir:            "/tmp/NotExist",
-			expectedResult: false,
-		},
-		{
-			desc:           "Existing dir",
-			dir:            existingMountPath,
-			expectedResult: false,
-		},
-	}
-
-	for i, test := range tests {
-		isCorruptedDir := IsCorruptedDir(test.dir)
-		assert.Equal(t, test.expectedResult, isCorruptedDir, "TestCase[%d]: %s", i, test.desc)
-	}
-}
 
 func TestCheckDiskCapacity(t *testing.T) {
 	d, _ := NewFakeDriver(t)
@@ -647,13 +80,13 @@ func TestRun(t *testing.T) {
 					}
 				}()
 
-				originalCredFile, ok := os.LookupEnv(DefaultAzureCredentialFileEnv)
+				originalCredFile, ok := os.LookupEnv(constants.DefaultAzureCredentialFileEnv)
 				if ok {
-					defer os.Setenv(DefaultAzureCredentialFileEnv, originalCredFile)
+					defer os.Setenv(constants.DefaultAzureCredentialFileEnv, originalCredFile)
 				} else {
-					defer os.Unsetenv(DefaultAzureCredentialFileEnv)
+					defer os.Unsetenv(constants.DefaultAzureCredentialFileEnv)
 				}
-				os.Setenv(DefaultAzureCredentialFileEnv, fakeCredFile)
+				os.Setenv(constants.DefaultAzureCredentialFileEnv, fakeCredFile)
 
 				d, _ := NewFakeDriver(t)
 				d.Run("tcp://127.0.0.1:0", "", true, true)
@@ -672,13 +105,13 @@ func TestRun(t *testing.T) {
 					}
 				}()
 
-				originalCredFile, ok := os.LookupEnv(DefaultAzureCredentialFileEnv)
+				originalCredFile, ok := os.LookupEnv(constants.DefaultAzureCredentialFileEnv)
 				if ok {
-					defer os.Setenv(DefaultAzureCredentialFileEnv, originalCredFile)
+					defer os.Setenv(constants.DefaultAzureCredentialFileEnv, originalCredFile)
 				} else {
-					defer os.Unsetenv(DefaultAzureCredentialFileEnv)
+					defer os.Unsetenv(constants.DefaultAzureCredentialFileEnv)
 				}
-				os.Setenv(DefaultAzureCredentialFileEnv, fakeCredFile)
+				os.Setenv(constants.DefaultAzureCredentialFileEnv, fakeCredFile)
 
 				d, _ := NewFakeDriver(t)
 				d.setCloud(&azure.Cloud{})
