+++ pkg/azuredisk/controllerserver.go
@@ -30,7 +30,6 @@ import (
 
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/types/known/timestamppb"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -39,42 +38,14 @@ import (
 	volerr "k8s.io/cloud-provider/volume/errors"
 	"k8s.io/klog/v2"
 
+	"sigs.k8s.io/azuredisk-csi-driver/pkg/azureutils"
 	"sigs.k8s.io/azuredisk-csi-driver/pkg/optimization"
 	volumehelper "sigs.k8s.io/azuredisk-csi-driver/pkg/util"
 	"sigs.k8s.io/cloud-provider-azure/pkg/metrics"
 	azure "sigs.k8s.io/cloud-provider-azure/pkg/provider"
 )
 
-var (
-	// volumeCaps represents how the volume could be accessed.
-	volumeCaps = []csi.VolumeCapability_AccessMode{
-		{
-			Mode: csi.VolumeCapability_AccessMode_SINGLE_NODE_WRITER,
-		},
-		{
-			Mode: csi.VolumeCapability_AccessMode_SINGLE_NODE_READER_ONLY,
-		},
-		{
-			Mode: csi.VolumeCapability_AccessMode_MULTI_NODE_READER_ONLY,
-		},
-		{
-			Mode: csi.VolumeCapability_AccessMode_MULTI_NODE_SINGLE_WRITER,
-		},
-		{
-			Mode: csi.VolumeCapability_AccessMode_MULTI_NODE_MULTI_WRITER,
-		},
-	}
-)
-
-const (
-	sourceSnapshot           = "snapshot"
-	sourceVolume             = "volume"
-	azureDiskCSIDriverName   = "azuredisk_csi_driver"
-	NotFound                 = "NotFound"
-	resizeRequired           = "resizeRequired"
-	requestedSizeGib         = "requestedsizegib"
-	sourceDiskSearchMaxDepth = 10
-)
+var ()
 
 // listVolumeStatus explains the return status of `listVolumesByResourceGroup`
 type listVolumeStatus struct {
@@ -108,8 +79,8 @@ func (d *Driver) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest)
 	capacityBytes := req.GetCapacityRange().GetRequiredBytes()
 	volSizeBytes := int64(capacityBytes)
 	requestGiB := int(volumehelper.RoundUpGiB(volSizeBytes))
-	if requestGiB < minimumDiskSizeGiB {
-		requestGiB = minimumDiskSizeGiB
+	if requestGiB < azureutils.MinimumDiskSizeGiB {
+		requestGiB = azureutils.MinimumDiskSizeGiB
 	}
 
 	maxVolSize := int(volumehelper.RoundUpGiB(req.GetCapacityRange().GetLimitBytes()))
@@ -143,34 +114,34 @@ func (d *Driver) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest)
 	}
 	for k, v := range parameters {
 		switch strings.ToLower(k) {
-		case skuNameField:
+		case azureutils.SkuNameField:
 			storageAccountType = v
-		case locationField:
+		case azureutils.LocationField:
 			location = v
-		case storageAccountTypeField:
+		case azureutils.StorageAccountTypeField:
 			storageAccountType = v
-		case cachingModeField:
+		case azureutils.CachingModeField:
 			cachingMode = v1.AzureDataDiskCachingMode(v)
-		case resourceGroupField:
+		case azureutils.ResourceGroupField:
 			resourceGroup = v
-		case diskIOPSReadWriteField:
+		case azureutils.DiskIOPSReadWriteField:
 			diskIopsReadWrite = v
-		case diskMBPSReadWriteField:
+		case azureutils.DiskMBPSReadWriteField:
 			diskMbpsReadWrite = v
-		case logicalSectorSizeField:
+		case azureutils.LogicalSectorSizeField:
 			logicalSectorSize, err = strconv.Atoi(v)
 			if err != nil {
 				return nil, status.Error(codes.InvalidArgument, fmt.Sprintf("parse %s failed with error: %v", v, err))
 			}
-		case diskNameField:
+		case azureutils.DiskNameField:
 			diskName = v
-		case desIDField:
+		case azureutils.DesIDField:
 			diskEncryptionSetID = v
-		case tagsField:
+		case azureutils.TagsField:
 			customTags = v
 		case azure.WriteAcceleratorEnabled:
 			writeAcceleratorEnabled = v
-		case maxSharesField:
+		case azureutils.MaxSharesField:
 			maxShares, err = strconv.Atoi(v)
 			if err != nil {
 				return nil, status.Error(codes.InvalidArgument, fmt.Sprintf("parse %s failed with error: %v", v, err))
@@ -178,27 +149,27 @@ func (d *Driver) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest)
 			if maxShares < 1 {
 				return nil, status.Error(codes.InvalidArgument, fmt.Sprintf("parse %s returned with invalid value: %d", v, maxShares))
 			}
-		case pvcNameKey:
-			tags[pvcNameTag] = v
-		case pvcNamespaceKey:
-			tags[pvcNamespaceTag] = v
-		case pvNameKey:
-			tags[pvNameTag] = v
-		case fsTypeField:
+		case azureutils.PvcNameKey:
+			tags[azureutils.PvcNameTag] = v
+		case azureutils.PvcNamespaceKey:
+			tags[azureutils.PvcNamespaceTag] = v
+		case azureutils.PvNameKey:
+			tags[azureutils.PvNameTag] = v
+		case azureutils.FsTypeField:
 			// no op, only used in NodeStageVolume
-		case kindField:
+		case azureutils.KindField:
 			// fix csi migration issue: https://github.com/kubernetes/kubernetes/issues/103433
-			parameters[kindField] = string(v1.AzureManagedDisk)
-		case perfProfileField:
+			parameters[azureutils.KindField] = string(v1.AzureManagedDisk)
+		case azureutils.PerfProfileField:
 			if !optimization.IsValidPerfProfile(v) {
 				return nil, status.Error(codes.InvalidArgument, fmt.Sprintf("Perf profile %s is not supported. Supported tuning modes are none and basic.", v))
 			}
-		case networkAccessPolicyField:
+		case azureutils.NetworkAccessPolicyField:
 			netAccessPolicy = v
-		case diskAccessIDField:
+		case azureutils.DiskAccessIDField:
 			diskAccessID = v
-		case enableBurstingField:
-			if strings.EqualFold(v, trueValue) {
+		case azureutils.EnableBurstingField:
+			if strings.EqualFold(v, azureutils.TrueValue) {
 				enableBursting = to.BoolPtr(true)
 			}
 		default:
@@ -206,7 +177,7 @@ func (d *Driver) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest)
 		}
 	}
 
-	if IsAzureStackCloud(d.cloud.Config.Cloud, d.cloud.Config.DisableAzureStackCloud) {
+	if azureutils.IsAzureStackCloud(d.cloud.Config.Cloud, d.cloud.Config.DisableAzureStackCloud) {
 		if maxShares > 1 {
 			return nil, status.Error(codes.InvalidArgument, fmt.Sprintf("Invalid maxShares value: %d as Azure Stack does not support shared disk.", maxShares))
 		}
@@ -225,29 +196,29 @@ func (d *Driver) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest)
 	if diskName == "" {
 		diskName = name
 	}
-	diskName = getValidDiskName(diskName)
+	diskName = azureutils.CreateValidDiskName(diskName)
 
 	if resourceGroup == "" {
 		resourceGroup = d.cloud.ResourceGroup
 	}
 
 	// normalize values
-	skuName, err := normalizeStorageAccountType(storageAccountType, d.cloud.Config.Cloud, d.cloud.Config.DisableAzureStackCloud)
+	skuName, err := azureutils.NormalizeStorageAccountType(storageAccountType, d.cloud.Config.Cloud, d.cloud.Config.DisableAzureStackCloud)
 	if err != nil {
 		return nil, err
 	}
 
-	if _, err = normalizeCachingMode(cachingMode); err != nil {
+	if _, err = azureutils.NormalizeCachingMode(cachingMode); err != nil {
 		return nil, err
 	}
 
-	networkAccessPolicy, err := normalizeNetworkAccessPolicy(netAccessPolicy)
+	networkAccessPolicy, err := azureutils.NormalizeNetworkAccessPolicy(netAccessPolicy)
 	if err != nil {
 		return nil, err
 	}
 
 	requirement := req.GetAccessibilityRequirements()
-	diskZone := pickAvailabilityZone(requirement, d.cloud.Location)
+	diskZone := azureutils.PickAvailabilityZone(requirement, d.cloud.Location, topologyKey)
 	accessibleTopology := []*csi.Topology{}
 	if skuName == compute.StandardSSDZRS || skuName == compute.PremiumZRS {
 		klog.V(2).Infof("diskZone(%s) is reset as empty since disk(%s) is ZRS(%s)", diskZone, diskName, skuName)
@@ -281,7 +252,7 @@ func (d *Driver) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest)
 		return nil, err
 	}
 
-	mc := metrics.NewMetricContext(azureDiskCSIDriverName, "controller_create_volume", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
+	mc := metrics.NewMetricContext(azureutils.AzureDiskCSIDriverName, "controller_create_volume", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
 	isOperationSucceeded := false
 	defer func() {
 		mc.ObserveOperationWithResult(isOperationSucceeded)
@@ -295,8 +266,8 @@ func (d *Driver) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest)
 		tags[k] = v
 	}
 
-	if strings.EqualFold(writeAcceleratorEnabled, trueValue) {
-		tags[azure.WriteAcceleratorEnabled] = trueValue
+	if strings.EqualFold(writeAcceleratorEnabled, azureutils.TrueValue) {
+		tags[azure.WriteAcceleratorEnabled] = azureutils.TrueValue
 	}
 	sourceID := ""
 	sourceType := ""
@@ -304,7 +275,7 @@ func (d *Driver) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest)
 	if content != nil {
 		if content.GetSnapshot() != nil {
 			sourceID = content.GetSnapshot().GetSnapshotId()
-			sourceType = sourceSnapshot
+			sourceType = azureutils.SourceSnapshot
 			contentSource = &csi.VolumeContentSource{
 				Type: &csi.VolumeContentSource_Snapshot{
 					Snapshot: &csi.VolumeContentSource_SnapshotSource{
@@ -314,7 +285,7 @@ func (d *Driver) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest)
 			}
 		} else {
 			sourceID = content.GetVolume().GetVolumeId()
-			sourceType = sourceVolume
+			sourceType = azureutils.SourceVolume
 			contentSource = &csi.VolumeContentSource{
 				Type: &csi.VolumeContentSource_Volume{
 					Volume: &csi.VolumeContentSource_VolumeSource{
@@ -324,14 +295,14 @@ func (d *Driver) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest)
 			}
 
 			ctx, cancel := context.WithCancel(context.Background())
-			if sourceGiB, _ := d.GetSourceDiskSize(ctx, resourceGroup, path.Base(sourceID), 0, sourceDiskSearchMaxDepth); sourceGiB != nil && *sourceGiB < int32(requestGiB) {
-				parameters[resizeRequired] = strconv.FormatBool(true)
+			if sourceGiB, _ := d.GetSourceDiskSize(ctx, resourceGroup, path.Base(sourceID), 0, azureutils.SourceDiskSearchMaxDepth); sourceGiB != nil && *sourceGiB < int32(requestGiB) {
+				parameters[azureutils.ResizeRequired] = strconv.FormatBool(true)
 			}
 			cancel()
 		}
 	}
 
-	parameters[requestedSizeGib] = strconv.Itoa(requestGiB)
+	parameters[azureutils.RequestedSizeGib] = strconv.Itoa(requestGiB)
 	volumeOptions := &azure.ManagedDiskOptions{
 		DiskName:            diskName,
 		StorageAccountType:  skuName,
@@ -351,7 +322,7 @@ func (d *Driver) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest)
 	}
 	volumeOptions.SkipGetDiskOperation = d.isGetDiskThrottled()
 	// Azure Stack Cloud does not support NetworkAccessPolicy
-	if !IsAzureStackCloud(d.cloud.Config.Cloud, d.cloud.Config.DisableAzureStackCloud) {
+	if !azureutils.IsAzureStackCloud(d.cloud.Config.Cloud, d.cloud.Config.DisableAzureStackCloud) {
 		volumeOptions.NetworkAccessPolicy = networkAccessPolicy
 		if diskAccessID != "" {
 			volumeOptions.DiskAccessID = &diskAccessID
@@ -359,7 +330,7 @@ func (d *Driver) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest)
 	}
 	diskURI, err := d.cloud.CreateManagedDisk(volumeOptions)
 	if err != nil {
-		if strings.Contains(err.Error(), NotFound) {
+		if strings.Contains(err.Error(), azureutils.NotFound) {
 			return nil, status.Error(codes.NotFound, err.Error())
 		}
 		return nil, err
@@ -391,7 +362,7 @@ func (d *Driver) DeleteVolume(ctx context.Context, req *csi.DeleteVolumeRequest)
 	}
 	diskURI := volumeID
 
-	if err := isValidDiskURI(diskURI); err != nil {
+	if err := azureutils.IsValidDiskURI(diskURI); err != nil {
 		klog.Errorf("validateDiskURI(%s) in DeleteVolume failed with error: %v", diskURI, err)
 		return &csi.DeleteVolumeResponse{}, nil
 	}
@@ -401,7 +372,7 @@ func (d *Driver) DeleteVolume(ctx context.Context, req *csi.DeleteVolumeRequest)
 	}
 	defer d.volumeLocks.Release(volumeID)
 
-	mc := metrics.NewMetricContext(azureDiskCSIDriverName, "controller_delete_volume", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
+	mc := metrics.NewMetricContext(azureutils.AzureDiskCSIDriverName, "controller_delete_volume", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
 	isOperationSucceeded := false
 	defer func() {
 		mc.ObserveOperationWithResult(isOperationSucceeded)
@@ -432,7 +403,7 @@ func (d *Driver) ControllerPublishVolume(ctx context.Context, req *csi.Controlle
 	}
 
 	caps := []*csi.VolumeCapability{volCap}
-	if !isValidVolumeCapabilities(caps) {
+	if !azureutils.IsValidVolumeCapabilities(caps) {
 		return nil, status.Error(codes.InvalidArgument, "Volume capability not supported")
 	}
 
@@ -447,12 +418,12 @@ func (d *Driver) ControllerPublishVolume(ctx context.Context, req *csi.Controlle
 	}
 
 	nodeName := types.NodeName(nodeID)
-	diskName, err := GetDiskName(diskURI)
+	diskName, err := azureutils.GetDiskName(diskURI)
 	if err != nil {
 		return nil, err
 	}
 
-	mc := metrics.NewMetricContext(azureDiskCSIDriverName, "controller_publish_volume", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
+	mc := metrics.NewMetricContext(azureutils.AzureDiskCSIDriverName, "controller_publish_volume", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
 	isOperationSucceeded := false
 	defer func() {
 		mc.ObserveOperationWithResult(isOperationSucceeded)
@@ -476,7 +447,7 @@ func (d *Driver) ControllerPublishVolume(ctx context.Context, req *csi.Controlle
 		klog.V(2).Infof("Attach operation is successful. volume %q is already attached to node %q at lun %d.", diskURI, nodeName, lun)
 	} else {
 		var cachingMode compute.CachingTypes
-		if cachingMode, err = getCachingMode(req.GetVolumeContext()); err != nil {
+		if cachingMode, err = azureutils.GetCachingMode(req.GetVolumeContext()); err != nil {
 			return nil, err
 		}
 		klog.V(2).Infof("Trying to attach volume %q to node %q", diskURI, nodeName)
@@ -506,7 +477,7 @@ func (d *Driver) ControllerPublishVolume(ctx context.Context, req *csi.Controlle
 		klog.V(2).Infof("attach volume %q to node %q successfully", diskURI, nodeName)
 	}
 
-	pvInfo := map[string]string{LUN: strconv.Itoa(int(lun))}
+	pvInfo := map[string]string{azureutils.LUN: strconv.Itoa(int(lun))}
 	isOperationSucceeded = true
 	return &csi.ControllerPublishVolumeResponse{PublishContext: pvInfo}, nil
 }
@@ -524,12 +495,12 @@ func (d *Driver) ControllerUnpublishVolume(ctx context.Context, req *csi.Control
 	}
 	nodeName := types.NodeName(nodeID)
 
-	diskName, err := GetDiskName(diskURI)
+	diskName, err := azureutils.GetDiskName(diskURI)
 	if err != nil {
 		return nil, err
 	}
 
-	mc := metrics.NewMetricContext(azureDiskCSIDriverName, "controller_unpublish_volume", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
+	mc := metrics.NewMetricContext(azureutils.AzureDiskCSIDriverName, "controller_unpublish_volume", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
 	isOperationSucceeded := false
 	defer func() {
 		mc.ObserveOperationWithResult(isOperationSucceeded)
@@ -538,7 +509,7 @@ func (d *Driver) ControllerUnpublishVolume(ctx context.Context, req *csi.Control
 	klog.V(2).Infof("Trying to detach volume %s from node %s", diskURI, nodeID)
 
 	if err := d.cloud.DetachDisk(diskName, diskURI, nodeName); err != nil {
-		if strings.Contains(err.Error(), errDiskNotFound) {
+		if strings.Contains(err.Error(), azureutils.ErrDiskNotFound) {
 			klog.Warningf("volume %s already detached from node %s", diskURI, nodeID)
 		} else {
 			return nil, status.Errorf(codes.Internal, "Could not detach volume %q from node %q: %v", diskURI, nodeID, err)
@@ -618,11 +589,11 @@ func (d *Driver) listVolumesInCluster(ctx context.Context, start, maxEntries int
 	for _, pv := range pvList.Items {
 		if pv.Spec.CSI != nil && pv.Spec.CSI.Driver == d.Name {
 			diskURI := pv.Spec.CSI.VolumeHandle
-			if err := isValidDiskURI(diskURI); err != nil {
+			if err := azureutils.IsValidDiskURI(diskURI); err != nil {
 				klog.Warningf("invalid disk uri (%s) with error(%v)", diskURI, err)
 				continue
 			}
-			rg, err := GetResourceGroupFromURI(diskURI)
+			rg, err := azureutils.GetResourceGroupFromURI(diskURI)
 			if err != nil {
 				klog.Warningf("failed to get resource group from disk uri (%s) with error(%v)", diskURI, err)
 				continue
@@ -790,20 +761,20 @@ func (d *Driver) ControllerExpandVolume(ctx context.Context, req *csi.Controller
 	requestSize := *resource.NewQuantity(capacityBytes, resource.BinarySI)
 
 	diskURI := req.GetVolumeId()
-	if err := isValidDiskURI(diskURI); err != nil {
+	if err := azureutils.IsValidDiskURI(diskURI); err != nil {
 		return nil, status.Errorf(codes.InvalidArgument, "disk URI(%s) is not valid: %v", diskURI, err)
 	}
 
-	diskName, err := GetDiskName(diskURI)
+	diskName, err := azureutils.GetDiskName(diskURI)
 	if err != nil {
 		return nil, status.Errorf(codes.Internal, "could not get disk name from diskURI(%s) with error(%v)", diskURI, err)
 	}
-	resourceGroup, err := GetResourceGroupFromURI(diskURI)
+	resourceGroup, err := azureutils.GetResourceGroupFromURI(diskURI)
 	if err != nil {
 		return nil, status.Errorf(codes.Internal, "could not get resource group from diskURI(%s) with error(%v)", diskURI, err)
 	}
 
-	mc := metrics.NewMetricContext(azureDiskCSIDriverName, "controller_expand_volume", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
+	mc := metrics.NewMetricContext(azureutils.AzureDiskCSIDriverName, "controller_expand_volume", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
 	isOperationSucceeded := false
 	defer func() {
 		mc.ObserveOperationWithResult(isOperationSucceeded)
@@ -849,7 +820,7 @@ func (d *Driver) CreateSnapshot(ctx context.Context, req *csi.CreateSnapshotRequ
 		return nil, status.Error(codes.InvalidArgument, "snapshot name must be provided")
 	}
 
-	snapshotName = getValidDiskName(snapshotName)
+	snapshotName = azureutils.CreateValidDiskName(snapshotName)
 
 	var customTags string
 	// set incremental snapshot as true by default
@@ -860,26 +831,26 @@ func (d *Driver) CreateSnapshot(ctx context.Context, req *csi.CreateSnapshotRequ
 	parameters := req.GetParameters()
 	for k, v := range parameters {
 		switch strings.ToLower(k) {
-		case tagsField:
+		case azureutils.TagsField:
 			customTags = v
-		case incrementalField:
+		case azureutils.IncrementalField:
 			if v == "false" {
 				incremental = false
 			}
-		case resourceGroupField:
+		case azureutils.ResourceGroupField:
 			resourceGroup = v
 		default:
 			return nil, fmt.Errorf("AzureDisk - invalid option %s in VolumeSnapshotClass", k)
 		}
 	}
 
-	if IsAzureStackCloud(d.cloud.Config.Cloud, d.cloud.Config.DisableAzureStackCloud) {
+	if azureutils.IsAzureStackCloud(d.cloud.Config.Cloud, d.cloud.Config.DisableAzureStackCloud) {
 		klog.V(2).Info("Use full snapshot instead as Azure Stack does not support incremental snapshot.")
 		incremental = false
 	}
 
 	if resourceGroup == "" {
-		resourceGroup, err = GetResourceGroupFromURI(sourceVolumeID)
+		resourceGroup, err = azureutils.GetResourceGroupFromURI(sourceVolumeID)
 		if err != nil {
 			return nil, status.Errorf(codes.InvalidArgument, "could not get resource group from diskURI(%s) with error(%v)", sourceVolumeID, err)
 		}
@@ -906,7 +877,7 @@ func (d *Driver) CreateSnapshot(ctx context.Context, req *csi.CreateSnapshotRequ
 		Tags:     tags,
 	}
 
-	mc := metrics.NewMetricContext(azureDiskCSIDriverName, "controller_create_snapshot", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
+	mc := metrics.NewMetricContext(azureutils.AzureDiskCSIDriverName, "controller_create_snapshot", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
 	isOperationSucceeded := false
 	defer func() {
 		mc.ObserveOperationWithResult(isOperationSucceeded)
@@ -946,14 +917,14 @@ func (d *Driver) DeleteSnapshot(ctx context.Context, req *csi.DeleteSnapshotRequ
 	snapshotName := snapshotID
 	resourceGroup := d.cloud.ResourceGroup
 
-	if isARMResourceID(snapshotID) {
+	if azureutils.IsARMResourceID(snapshotID) {
 		snapshotName, resourceGroup, err = d.getSnapshotInfo(snapshotID)
 		if err != nil {
 			return nil, err
 		}
 	}
 
-	mc := metrics.NewMetricContext(azureDiskCSIDriverName, "controller_delete_snapshot", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
+	mc := metrics.NewMetricContext(azureutils.AzureDiskCSIDriverName, "controller_delete_snapshot", d.cloud.ResourceGroup, d.cloud.SubscriptionID, d.Name)
 	isOperationSucceeded := false
 	defer func() {
 		mc.ObserveOperationWithResult(isOperationSucceeded)
@@ -975,7 +946,7 @@ func (d *Driver) ListSnapshots(ctx context.Context, req *csi.ListSnapshotsReques
 	if len(req.GetSnapshotId()) != 0 {
 		snapshot, err := d.getSnapshotByID(ctx, d.cloud.ResourceGroup, req.GetSnapshotId(), req.SourceVolumeId)
 		if err != nil {
-			if strings.Contains(err.Error(), resourceNotFound) {
+			if strings.Contains(err.Error(), azureutils.ResourceNotFound) {
 				return &csi.ListSnapshotsResponse{}, nil
 			}
 			return nil, err
@@ -997,13 +968,13 @@ func (d *Driver) ListSnapshots(ctx context.Context, req *csi.ListSnapshotsReques
 		return nil, status.Error(codes.Internal, fmt.Sprintf("Unknown list snapshot error: %v", err.Error()))
 	}
 
-	return getEntriesAndNextToken(req, snapshots)
+	return azureutils.GetEntriesAndNextToken(req, snapshots)
 }
 
 func (d *Driver) getSnapshotByID(ctx context.Context, resourceGroup, snapshotID, sourceVolumeID string) (*csi.Snapshot, error) {
 	var err error
 	snapshotName := snapshotID
-	if isARMResourceID(snapshotID) {
+	if azureutils.IsARMResourceID(snapshotID) {
 		snapshotName, resourceGroup, err = d.getSnapshotInfo(snapshotID)
 		if err != nil {
 			return nil, err
@@ -1015,7 +986,7 @@ func (d *Driver) getSnapshotByID(ctx context.Context, resourceGroup, snapshotID,
 		return nil, status.Error(codes.Internal, fmt.Sprintf("get snapshot %s from rg(%s) error: %v", snapshotName, resourceGroup, rerr.Error()))
 	}
 
-	return generateCSISnapshot(sourceVolumeID, &snapshot)
+	return azureutils.GenerateCSISnapshot(sourceVolumeID, &snapshot)
 }
 
 // GetSourceDiskSize recursively searches for the sourceDisk and returns: sourceDisk disk size, error
@@ -1034,7 +1005,7 @@ func (d *Driver) GetSourceDiskSize(ctx context.Context, resourceGroup, diskName 
 	if result.DiskProperties.CreationData != nil && (*result.DiskProperties.CreationData).CreateOption == "Copy" {
 		klog.V(2).Infof("Clone source disk has a parent source")
 		sourceResourceID := *result.DiskProperties.CreationData.SourceResourceID
-		parentResourceGroup, _ := GetResourceGroupFromURI(sourceResourceID)
+		parentResourceGroup, _ := azureutils.GetResourceGroupFromURI(sourceResourceID)
 		parentDiskName := path.Base(sourceResourceID)
 		return d.GetSourceDiskSize(ctx, parentResourceGroup, parentDiskName, curDepth+1, maxDepth)
 	}
@@ -1045,180 +1016,13 @@ func (d *Driver) GetSourceDiskSize(ctx context.Context, resourceGroup, diskName 
 	return (*result.DiskProperties).DiskSizeGB, nil
 }
 
-func isValidVolumeCapabilities(volCaps []*csi.VolumeCapability) bool {
-	hasSupport := func(cap *csi.VolumeCapability) bool {
-		for _, c := range volumeCaps {
-			// todo: Block volume support
-			/* compile error here
-			if blk := c.GetBlock(); blk != nil {
-				return false
-			}
-			*/
-			if c.GetMode() == cap.AccessMode.GetMode() {
-				return true
-			}
-		}
-		return false
-	}
-
-	foundAll := true
-	for _, c := range volCaps {
-		if !hasSupport(c) {
-			foundAll = false
-		}
-	}
-	return foundAll
-}
-
-// pickAvailabilityZone selects 1 zone given topology requirement.
-// if not found or topology requirement is not zone format, empty string is returned.
-func pickAvailabilityZone(requirement *csi.TopologyRequirement, region string) string {
-	if requirement == nil {
-		return ""
-	}
-	for _, topology := range requirement.GetPreferred() {
-		if zone, exists := topology.GetSegments()[WellKnownTopologyKey]; exists {
-			if isAvailabilityZone(zone, region) {
-				return zone
-			}
-		}
-		if zone, exists := topology.GetSegments()[topologyKey]; exists {
-			if isAvailabilityZone(zone, region) {
-				return zone
-			}
-		}
-	}
-	for _, topology := range requirement.GetRequisite() {
-		if zone, exists := topology.GetSegments()[WellKnownTopologyKey]; exists {
-			if isAvailabilityZone(zone, region) {
-				return zone
-			}
-		}
-		if zone, exists := topology.GetSegments()[topologyKey]; exists {
-			if isAvailabilityZone(zone, region) {
-				return zone
-			}
-		}
-	}
-	return ""
-}
-
-func getCachingMode(attributes map[string]string) (compute.CachingTypes, error) {
-	var (
-		cachingMode v1.AzureDataDiskCachingMode
-		err         error
-	)
-
-	for k, v := range attributes {
-		switch strings.ToLower(k) {
-		case cachingModeField:
-			cachingMode = v1.AzureDataDiskCachingMode(v)
-		}
-	}
-
-	cachingMode, err = normalizeCachingMode(cachingMode)
-	return compute.CachingTypes(cachingMode), err
-}
-
-func generateCSISnapshot(sourceVolumeID string, snapshot *compute.Snapshot) (*csi.Snapshot, error) {
-	if snapshot == nil || snapshot.SnapshotProperties == nil {
-		return nil, fmt.Errorf("snapshot property is nil")
-	}
-
-	tp := timestamppb.New(snapshot.SnapshotProperties.TimeCreated.ToTime())
-	if tp == nil {
-		return nil, fmt.Errorf("Failed to covert timestamp(%v)", snapshot.SnapshotProperties.TimeCreated.ToTime())
-	}
-	ready, _ := isCSISnapshotReady(*snapshot.SnapshotProperties.ProvisioningState)
-
-	if snapshot.SnapshotProperties.DiskSizeGB == nil {
-		return nil, fmt.Errorf("diskSizeGB of snapshot property is nil")
-	}
-
-	if sourceVolumeID == "" {
-		sourceVolumeID = getSourceVolumeID(snapshot)
-	}
-
-	return &csi.Snapshot{
-		SizeBytes:      volumehelper.GiBToBytes(int64(*snapshot.SnapshotProperties.DiskSizeGB)),
-		SnapshotId:     *snapshot.ID,
-		SourceVolumeId: sourceVolumeID,
-		CreationTime:   tp,
-		ReadyToUse:     ready,
-	}, nil
-}
-
-func isCSISnapshotReady(state string) (bool, error) {
-	switch strings.ToLower(state) {
-	case "succeeded":
-		return true, nil
-	default:
-		return false, nil
-	}
-}
-
 // The format of snapshot id is /subscriptions/xxx/resourceGroups/xxx/providers/Microsoft.Compute/snapshots/snapshot-xxx-xxx.
 func (d *Driver) getSnapshotInfo(snapshotID string) (snapshotName, resourceGroup string, err error) {
-	if snapshotName, err = getSnapshotName(snapshotID); err != nil {
+	if snapshotName, err = azureutils.GetSnapshotNameFromURI(snapshotID); err != nil {
 		return "", "", err
 	}
-	if resourceGroup, err = GetResourceGroupFromURI(snapshotID); err != nil {
+	if resourceGroup, err = azureutils.GetResourceGroupFromURI(snapshotID); err != nil {
 		return "", "", err
 	}
 	return snapshotName, resourceGroup, err
-}
-
-// There are 4 scenarios for listing snapshots.
-// 1. StartingToken is null, and MaxEntries is null. Return all snapshots from zero.
-// 2. StartingToken is null, and MaxEntries is not null. Return `MaxEntries` snapshots from zero.
-// 3. StartingToken is not null, and MaxEntries is null. Return all snapshots from `StartingToken`.
-// 4. StartingToken is not null, and MaxEntries is not null. Return `MaxEntries` snapshots from `StartingToken`.
-func getEntriesAndNextToken(req *csi.ListSnapshotsRequest, snapshots []compute.Snapshot) (*csi.ListSnapshotsResponse, error) {
-	if req == nil {
-		return nil, status.Errorf(codes.Aborted, "request is nil")
-	}
-
-	var err error
-	start := 0
-	if req.StartingToken != "" {
-		start, err = strconv.Atoi(req.StartingToken)
-		if err != nil {
-			return nil, status.Errorf(codes.Aborted, "ListSnapshots starting token(%s) parsing with error: %v", req.StartingToken, err)
-
-		}
-		if start >= len(snapshots) {
-			return nil, status.Errorf(codes.Aborted, "ListSnapshots starting token(%d) is greater than total number of snapshots", start)
-		}
-		if start < 0 {
-			return nil, status.Errorf(codes.Aborted, "ListSnapshots starting token(%d) can not be negative", start)
-		}
-	}
-
-	maxEntries := len(snapshots) - start
-	if req.MaxEntries > 0 && int(req.MaxEntries) < maxEntries {
-		maxEntries = int(req.MaxEntries)
-	}
-	entries := []*csi.ListSnapshotsResponse_Entry{}
-	for count := 0; start < len(snapshots) && count < maxEntries; start++ {
-		if (req.SourceVolumeId != "" && req.SourceVolumeId == getSourceVolumeID(&snapshots[start])) || req.SourceVolumeId == "" {
-			csiSnapshot, err := generateCSISnapshot(req.SourceVolumeId, &snapshots[start])
-			if err != nil {
-				return nil, fmt.Errorf("failed to generate snapshot entry: %v", err)
-			}
-			entries = append(entries, &csi.ListSnapshotsResponse_Entry{Snapshot: csiSnapshot})
-			count++
-		}
-	}
-
-	nextToken := len(snapshots)
-	if start < len(snapshots) {
-		nextToken = start
-	}
-
-	listSnapshotResp := &csi.ListSnapshotsResponse{
-		Entries:   entries,
-		NextToken: strconv.Itoa(nextToken),
-	}
-
-	return listSnapshotResp, nil
 }
