+++ pkg/azuredisk/nodeserver.go
@@ -38,6 +38,7 @@ import (
 	"k8s.io/kubernetes/pkg/volume"
 	"k8s.io/kubernetes/pkg/volume/util/hostutil"
 	mount "k8s.io/mount-utils"
+	"sigs.k8s.io/azuredisk-csi-driver/pkg/azureutils"
 )
 
 const (
@@ -72,7 +73,7 @@ func (d *Driver) NodeStageVolume(ctx context.Context, req *csi.NodeStageVolumeRe
 		return nil, status.Error(codes.InvalidArgument, "Volume capability not provided")
 	}
 
-	if !isValidVolumeCapabilities([]*csi.VolumeCapability{volumeCapability}) {
+	if !azureutils.IsValidVolumeCapabilities([]*csi.VolumeCapability{volumeCapability}) {
 		return nil, status.Error(codes.InvalidArgument, "Volume capability not supported")
 	}
 
@@ -81,7 +82,7 @@ func (d *Driver) NodeStageVolume(ctx context.Context, req *csi.NodeStageVolumeRe
 	}
 	defer d.volumeLocks.Release(diskURI)
 
-	lun, ok := req.PublishContext[LUN]
+	lun, ok := req.PublishContext[azureutils.LUN]
 	if !ok {
 		return nil, status.Error(codes.InvalidArgument, "lun not provided")
 	}
@@ -141,7 +142,7 @@ func (d *Driver) NodeStageVolume(ctx context.Context, req *csi.NodeStageVolumeRe
 	}
 
 	// If partition is specified, should mount it only instead of the entire disk.
-	if partition, ok := req.GetVolumeContext()[volumeAttributePartition]; ok {
+	if partition, ok := req.GetVolumeContext()[azureutils.VolumeAttributePartition]; ok {
 		source = source + "-part" + partition
 	}
 
@@ -155,7 +156,7 @@ func (d *Driver) NodeStageVolume(ctx context.Context, req *csi.NodeStageVolumeRe
 	klog.V(2).Infof("NodeStageVolume: format %s and mounting at %s successfully.", source, target)
 
 	// if resize is required, resize filesystem
-	if required, ok := req.GetVolumeContext()[resizeRequired]; ok && required == "true" {
+	if required, ok := req.GetVolumeContext()[azureutils.ResizeRequired]; ok && required == "true" {
 		klog.V(2).Infof("NodeStageVolume: fs resize initiating on target(%s) volumeid(%s)", target, diskURI)
 		if err != nil {
 			return nil, status.Errorf(codes.Internal, "NodeStageVolume: Could not get volume path for %s: %v", target, err)
@@ -231,7 +232,7 @@ func (d *Driver) NodePublishVolume(ctx context.Context, req *csi.NodePublishVolu
 
 	switch req.GetVolumeCapability().GetAccessType().(type) {
 	case *csi.VolumeCapability_Block:
-		lun, ok := req.PublishContext[LUN]
+		lun, ok := req.PublishContext[azureutils.LUN]
 		if !ok {
 			return nil, status.Error(codes.InvalidArgument, "lun not provided")
 		}
@@ -316,9 +317,9 @@ func (d *Driver) NodeGetInfo(ctx context.Context, req *csi.NodeGetInfoRequest) (
 	if err != nil {
 		klog.Warningf("Failed to get zone from Azure cloud provider, nodeName: %v, error: %v", d.NodeID, err)
 	} else {
-		if isAvailabilityZone(zone.FailureDomain, d.cloud.Location) {
+		if azureutils.IsValidAvailabilityZone(zone.FailureDomain, d.cloud.Location) {
 			topology.Segments[topologyKey] = zone.FailureDomain
-			topology.Segments[WellKnownTopologyKey] = zone.FailureDomain
+			topology.Segments[azureutils.WellKnownTopologyKey] = zone.FailureDomain
 			klog.V(2).Infof("NodeGetInfo, nodeName: %v, zone: %v", d.NodeID, zone.FailureDomain)
 		}
 	}
@@ -460,7 +461,7 @@ func (d *Driver) NodeExpandVolume(ctx context.Context, req *csi.NodeExpandVolume
 	if volumeCapability != nil {
 		// VolumeCapability is optional, if specified, validate it
 		caps := []*csi.VolumeCapability{volumeCapability}
-		if !isValidVolumeCapabilities(caps) {
+		if !azureutils.IsValidVolumeCapabilities(caps) {
 			return nil, status.Error(codes.InvalidArgument, "VolumeCapability is invalid.")
 		}
 
@@ -500,7 +501,7 @@ func (d *Driver) NodeExpandVolume(ctx context.Context, req *csi.NodeExpandVolume
 func getFStype(attributes map[string]string) string {
 	for k, v := range attributes {
 		switch strings.ToLower(k) {
-		case fsTypeField:
+		case azureutils.FsTypeField:
 			return strings.ToLower(v)
 		}
 	}
@@ -513,7 +514,7 @@ func getFStype(attributes map[string]string) string {
 func (d *Driver) ensureMountPoint(target string) (bool, error) {
 	notMnt, err := d.mounter.IsLikelyNotMountPoint(target)
 	if err != nil && !os.IsNotExist(err) {
-		if IsCorruptedDir(target) {
+		if azureutils.IsCorruptedDir(target) {
 			notMnt = false
 			klog.Warningf("detected corrupted mount for targetPath [%s]", target)
 		} else {
@@ -554,7 +555,7 @@ func (d *Driver) formatAndMount(source, target, fstype string, options []string)
 }
 
 func (d *Driver) getDevicePathWithLUN(lunStr string) (string, error) {
-	lun, err := getDiskLUN(lunStr)
+	lun, err := azureutils.GetDiskLUN(lunStr)
 	if err != nil {
 		return "", err
 	}
