+++ pkg/azuredisk/nodeserver_test.go
@@ -34,6 +34,7 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
+	"sigs.k8s.io/azuredisk-csi-driver/pkg/azureutils"
 	volumehelper "sigs.k8s.io/azuredisk-csi-driver/pkg/util"
 )
 
@@ -286,7 +287,7 @@ func TestNodeStageVolume(t *testing.T) {
 	volumeCap := csi.VolumeCapability_AccessMode{Mode: 2}
 	volumeCapWrong := csi.VolumeCapability_AccessMode{Mode: 10}
 	publishContext := map[string]string{
-		LUN: "/dev/01",
+		azureutils.LUN: "/dev/01",
 	}
 
 	tests := []struct {
@@ -472,18 +473,18 @@ func TestNodePublishVolume(t *testing.T) {
 
 	volumeCap := csi.VolumeCapability_AccessMode{Mode: csi.VolumeCapability_AccessMode_MULTI_NODE_MULTI_WRITER}
 	publishContext := map[string]string{
-		LUN: "/dev/01",
+		azureutils.LUN: "/dev/01",
 	}
 	errorMountSource, err := testutil.GetWorkDirPath("error_mount_source")
 	assert.NoError(t, err)
 	alreadyMountedTarget, err := testutil.GetWorkDirPath("false_is_likely_exist_target")
 	assert.NoError(t, err)
 
-	azurediskPath := "azure.go"
+	azurediskPath := "azuredisk.go"
 
 	// ".\azure.go will get deleted on Windows"
 	if runtime.GOOS == "windows" {
-		azurediskPath = "testfiles\\azure.go"
+		azurediskPath = "testfiles\\azuredisk.go"
 	}
 	azuredisk, err := testutil.GetWorkDirPath(azurediskPath)
 	assert.NoError(t, err)
@@ -925,7 +926,7 @@ func TestMakeDir(t *testing.T) {
 	assert.NoError(t, err)
 
 	//Failed case
-	err = makeDir("./azure.go")
+	err = makeDir("./azuredisk.go")
 	var e *os.PathError
 	if !errors.As(err, &e) {
 		t.Errorf("Unexpected Error: %v", err)
