+++ pkg/azurediskplugin/main.go
@@ -29,6 +29,7 @@ import (
 
 	"k8s.io/component-base/metrics/legacyregistry"
 	"k8s.io/klog/v2"
+	constants "sigs.k8s.io/azuredisk-csi-driver/pkg/azureutils"
 )
 
 func init() {
@@ -41,7 +42,7 @@ var (
 	version                    = flag.Bool("version", false, "Print the version and exit.")
 	metricsAddress             = flag.String("metrics-address", "0.0.0.0:29604", "export the metrics")
 	kubeconfig                 = flag.String("kubeconfig", "", "Absolute path to the kubeconfig file. Required only when running out of cluster.")
-	driverName                 = flag.String("drivername", azuredisk.DefaultDriverName, "name of the driver")
+	driverName                 = flag.String("drivername", constants.DefaultDriverName, "name of the driver")
 	volumeAttachLimit          = flag.Int64("volume-attach-limit", -1, "maximum number of attachable volumes per node")
 	disableAVSetNodes          = flag.Bool("disable-avset-nodes", true, "disable DisableAvailabilitySetNodes in cloud config for controller")
 	enablePerfOptimization     = flag.Bool("enable-perf-optimization", false, "boolean flag to enable disk perf optimization")
