+++ pkg/azureutils/azure_disk_utils.go
@@ -1,0 +1,431 @@
+/*
+Copyright 2021 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package azureutils
+
+import (
+	"fmt"
+	"os"
+	"regexp"
+	"strconv"
+	"strings"
+	"unicode"
+
+	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2020-12-01/compute"
+	"github.com/container-storage-interface/spec/lib/go/csi"
+	"github.com/pborman/uuid"
+	v1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/util/sets"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
+	"k8s.io/klog/v2"
+	"k8s.io/mount-utils"
+
+	api "k8s.io/kubernetes/pkg/apis/core"
+	volumeUtil "k8s.io/kubernetes/pkg/volume/util"
+	"sigs.k8s.io/azuredisk-csi-driver/pkg/util"
+	azure "sigs.k8s.io/cloud-provider-azure/pkg/provider"
+)
+
+const (
+	azurePublicCloud                          = "AZUREPUBLICCLOUD"
+	azureStackCloud                           = "AZURESTACKCLOUD"
+	azurePublicCloudDefaultStorageAccountType = compute.StandardSSDLRS
+	azureStackCloudDefaultStorageAccountType  = compute.StandardLRS
+	defaultAzureDataDiskCachingMode           = v1.AzureDataDiskCachingReadOnly
+	// default IOPS Caps & Throughput Cap (MBps) per https://docs.microsoft.com/en-us/azure/virtual-machines/linux/disks-ultra-ssd
+	// see https://docs.microsoft.com/en-us/rest/api/compute/disks/createorupdate#uri-parameters
+	diskNameMinLength         = 1
+	diskNameMaxLength         = 80
+	diskNameGenerateMaxLength = 76 // maxLength = 80 - (4 for ".vhd") = 76
+)
+
+var (
+	// see https://docs.microsoft.com/en-us/rest/api/compute/disks/createorupdate#create-a-managed-disk-by-copying-a-snapshot.
+	diskSnapshotPath        = "/subscriptions/%s/resourceGroups/%s/providers/Microsoft.Compute/snapshots/%s"
+	diskSnapshotPathRE      = regexp.MustCompile(`(?i).*/subscriptions/(?:.*)/resourceGroups/(?:.*)/providers/Microsoft.Compute/snapshots/(.+)`)
+	diskURISupportedManaged = []string{"/subscriptions/{sub-id}/resourcegroups/{group-name}/providers/microsoft.compute/disks/{disk-id}"}
+	lunPathRE               = regexp.MustCompile(`/dev(?:.*)/disk/azure/scsi(?:.*)/lun(.+)`)
+	supportedCachingModes   = sets.NewString(
+		string(api.AzureDataDiskCachingNone),
+		string(api.AzureDataDiskCachingReadOnly),
+		string(api.AzureDataDiskCachingReadWrite),
+	)
+
+	// volumeCaps represents how the volume could be accessed.
+	volumeCaps = []csi.VolumeCapability_AccessMode{
+		{
+			Mode: csi.VolumeCapability_AccessMode_SINGLE_NODE_WRITER,
+		},
+		{
+			Mode: csi.VolumeCapability_AccessMode_SINGLE_NODE_READER_ONLY,
+		},
+		{
+			Mode: csi.VolumeCapability_AccessMode_MULTI_NODE_READER_ONLY,
+		},
+		{
+			Mode: csi.VolumeCapability_AccessMode_MULTI_NODE_SINGLE_WRITER,
+		},
+		{
+			Mode: csi.VolumeCapability_AccessMode_MULTI_NODE_MULTI_WRITER,
+		},
+	}
+)
+
+func GetCachingMode(attributes map[string]string) (compute.CachingTypes, error) {
+	var (
+		cachingMode v1.AzureDataDiskCachingMode
+		err         error
+	)
+
+	for k, v := range attributes {
+		if strings.EqualFold(k, CachingModeField) {
+			cachingMode = v1.AzureDataDiskCachingMode(v)
+			break
+		}
+	}
+
+	cachingMode, err = NormalizeCachingMode(cachingMode)
+	return compute.CachingTypes(cachingMode), err
+}
+
+// GetCloudProvider get Azure Cloud Provider
+func GetCloudProvider(kubeconfig, secretName, secretNamespace, userAgent string) (*azure.Cloud, error) {
+	kubeClient, err := getKubeClient(kubeconfig)
+	if err != nil {
+		klog.Warningf("get kubeconfig(%s) failed with error: %v", kubeconfig, err)
+		if !os.IsNotExist(err) && err != rest.ErrNotInCluster {
+			return nil, fmt.Errorf("failed to get KubeClient: %v", err)
+		}
+	}
+
+	az := &azure.Cloud{
+		InitSecretConfig: azure.InitSecretConfig{
+			SecretName:      secretName,
+			SecretNamespace: secretNamespace,
+			CloudConfigKey:  "cloud-config",
+		},
+	}
+	if kubeClient != nil {
+		klog.V(2).Infof("reading cloud config from secret")
+		az.KubeClient = kubeClient
+		if err := az.InitializeCloudFromSecret(); err != nil {
+			klog.V(2).Infof("InitializeCloudFromSecret failed with error: %v", err)
+		}
+	}
+
+	if az.TenantID == "" || az.SubscriptionID == "" || az.ResourceGroup == "" {
+		klog.V(2).Infof("could not read cloud config from secret")
+		credFile, ok := os.LookupEnv(DefaultAzureCredentialFileEnv)
+		if ok && strings.TrimSpace(credFile) != "" {
+			klog.V(2).Infof("%s env var set as %v", DefaultAzureCredentialFileEnv, credFile)
+		} else {
+			if util.IsWindowsOS() {
+				credFile = DefaultCredFilePathWindows
+			} else {
+				credFile = DefaultCredFilePathLinux
+			}
+
+			klog.V(2).Infof("use default %s env var: %v", DefaultAzureCredentialFileEnv, credFile)
+		}
+
+		var config *os.File
+		config, err = os.Open(credFile)
+		if err != nil {
+			klog.Errorf("load azure config from file(%s) failed with %v", credFile, err)
+			return nil, fmt.Errorf("load azure config from file(%s) failed with %v", credFile, err)
+		}
+		defer config.Close()
+
+		klog.V(2).Infof("read cloud config from file: %s successfully", credFile)
+		if az, err = azure.NewCloudWithoutFeatureGates(config, false); err != nil {
+			return az, err
+		}
+	}
+
+	// reassign kubeClient
+	if kubeClient != nil && az.KubeClient == nil {
+		az.KubeClient = kubeClient
+	}
+	return az, nil
+}
+
+// GetDiskLUN : deviceInfo could be a LUN number or a device path, e.g. /dev/disk/azure/scsi1/lun2
+func GetDiskLUN(deviceInfo string) (int32, error) {
+	var diskLUN string
+	if len(deviceInfo) <= 2 {
+		diskLUN = deviceInfo
+	} else {
+		// extract the LUN num from a device path
+		matches := lunPathRE.FindStringSubmatch(deviceInfo)
+		if len(matches) == 2 {
+			diskLUN = matches[1]
+		} else {
+			return -1, fmt.Errorf("cannot parse deviceInfo: %s", deviceInfo)
+		}
+	}
+
+	lun, err := strconv.Atoi(diskLUN)
+	if err != nil {
+		return -1, err
+	}
+	return int32(lun), nil
+}
+
+func GetDiskName(diskURI string) (string, error) {
+	matches := ManagedDiskPathRE.FindStringSubmatch(diskURI)
+	if len(matches) != 2 {
+		return "", fmt.Errorf("could not get disk name from %s, correct format: %s", diskURI, ManagedDiskPathRE)
+	}
+	return matches[1], nil
+}
+
+// Disk name must begin with a letter or number, end with a letter, number or underscore,
+// and may contain only letters, numbers, underscores, periods, or hyphens.
+// See https://docs.microsoft.com/en-us/rest/api/compute/disks/createorupdate#uri-parameters
+//
+//
+// Snapshot name must begin with a letter or number, end with a letter, number or underscore,
+// and may contain only letters, numbers, underscores, periods, or hyphens.
+// See https://docs.microsoft.com/en-us/rest/api/compute/snapshots/createorupdate#uri-parameters
+//
+// Since the naming rule of disk is same with snapshot's, here we use the same function to handle disks and snapshots.
+func CreateValidDiskName(volumeName string) string {
+	diskName := volumeName
+	if len(diskName) > diskNameMaxLength {
+		diskName = diskName[0:diskNameMaxLength]
+		klog.Warningf("since the maximum volume name length is %d, so it is truncated as (%q)", diskNameMaxLength, diskName)
+	}
+	if !checkDiskName(diskName) || len(diskName) < diskNameMinLength {
+		// todo: get cluster name
+		diskName = volumeUtil.GenerateVolumeName("pvc-disk", uuid.NewUUID().String(), diskNameGenerateMaxLength)
+		klog.Warningf("the requested volume name (%q) is invalid, so it is regenerated as (%q)", volumeName, diskName)
+	}
+
+	return diskName
+}
+
+func GetResourceGroupFromURI(diskURI string) (string, error) {
+	fields := strings.Split(diskURI, "/")
+	if len(fields) != 9 || strings.ToLower(fields[3]) != "resourcegroups" {
+		return "", fmt.Errorf("invalid disk URI: %s", diskURI)
+	}
+	return fields[4], nil
+}
+
+func GetValidCreationData(subscriptionID, resourceGroup, sourceResourceID, sourceType string) (compute.CreationData, error) {
+	if sourceResourceID == "" {
+		return compute.CreationData{
+			CreateOption: compute.Empty,
+		}, nil
+	}
+
+	switch sourceType {
+	case SourceSnapshot:
+		if match := diskSnapshotPathRE.FindString(sourceResourceID); match == "" {
+			sourceResourceID = fmt.Sprintf(diskSnapshotPath, subscriptionID, resourceGroup, sourceResourceID)
+		}
+
+	case SourceVolume:
+		if match := ManagedDiskPathRE.FindString(sourceResourceID); match == "" {
+			sourceResourceID = fmt.Sprintf(ManagedDiskPath, subscriptionID, resourceGroup, sourceResourceID)
+		}
+	default:
+		return compute.CreationData{
+			CreateOption: compute.Empty,
+		}, nil
+	}
+
+	splits := strings.Split(sourceResourceID, "/")
+	if len(splits) > 9 {
+		if sourceType == SourceSnapshot {
+			return compute.CreationData{}, fmt.Errorf("sourceResourceID(%s) is invalid, correct format: %s", sourceResourceID, diskSnapshotPathRE)
+		}
+
+		return compute.CreationData{}, fmt.Errorf("sourceResourceID(%s) is invalid, correct format: %s", sourceResourceID, ManagedDiskPathRE)
+	}
+	return compute.CreationData{
+		CreateOption:     compute.Copy,
+		SourceResourceID: &sourceResourceID,
+	}, nil
+}
+
+func IsCorruptedDir(dir string) bool {
+	_, pathErr := mount.PathExists(dir)
+	fmt.Printf("IsCorruptedDir(%s) returned with error: %v", dir, pathErr)
+	return pathErr != nil && mount.IsCorruptedMnt(pathErr)
+}
+
+// IsARMResourceID check whether resourceID is an ARM ResourceID
+func IsARMResourceID(resourceID string) bool {
+	id := strings.ToLower(resourceID)
+	return strings.Contains(id, "/subscriptions/")
+}
+
+// IsAzureStackCloud decides whether the driver is running on Azure Stack Cloud.
+func IsAzureStackCloud(cloud string, disableAzureStackCloud bool) bool {
+	return !disableAzureStackCloud && strings.EqualFold(cloud, azureStackCloud)
+}
+
+// IsValidAvailabilityZone returns true if the zone is in format of <region>-<zone-id>.
+func IsValidAvailabilityZone(zone, region string) bool {
+	return strings.HasPrefix(zone, fmt.Sprintf("%s-", region))
+}
+
+func IsValidDiskURI(diskURI string) error {
+	if strings.Index(strings.ToLower(diskURI), "/subscriptions/") != 0 {
+		return fmt.Errorf("inavlid DiskURI: %v, correct format: %v", diskURI, diskURISupportedManaged)
+	}
+	return nil
+}
+
+func IsValidVolumeCapabilities(volCaps []*csi.VolumeCapability) bool {
+	hasSupport := func(cap *csi.VolumeCapability) bool {
+		for _, c := range volumeCaps {
+			// todo: Block volume support
+			/* compile error here
+			if blk := c.GetBlock(); blk != nil {
+				return false
+			}
+			*/
+			if c.GetMode() == cap.AccessMode.GetMode() {
+				return true
+			}
+		}
+		return false
+	}
+
+	foundAll := true
+	for _, c := range volCaps {
+		if !hasSupport(c) {
+			foundAll = false
+		}
+	}
+	return foundAll
+}
+
+func NormalizeCachingMode(cachingMode v1.AzureDataDiskCachingMode) (v1.AzureDataDiskCachingMode, error) {
+	if cachingMode == "" {
+		return defaultAzureDataDiskCachingMode, nil
+	}
+
+	if !supportedCachingModes.Has(string(cachingMode)) {
+		return "", fmt.Errorf("azureDisk - %s is not supported cachingmode. Supported values are %s", cachingMode, supportedCachingModes.List())
+	}
+
+	return cachingMode, nil
+}
+
+func NormalizeNetworkAccessPolicy(networkAccessPolicy string) (compute.NetworkAccessPolicy, error) {
+	if networkAccessPolicy == "" {
+		return compute.AllowAll, nil
+	}
+	policy := compute.NetworkAccessPolicy(networkAccessPolicy)
+	for _, s := range compute.PossibleNetworkAccessPolicyValues() {
+		if policy == s {
+			return policy, nil
+		}
+	}
+	return "", fmt.Errorf("azureDisk - %s is not supported NetworkAccessPolicy. Supported values are %s", networkAccessPolicy, compute.PossibleNetworkAccessPolicyValues())
+}
+
+func NormalizeStorageAccountType(storageAccountType, cloud string, disableAzureStackCloud bool) (compute.DiskStorageAccountTypes, error) {
+	if storageAccountType == "" {
+		if IsAzureStackCloud(cloud, disableAzureStackCloud) {
+			return azureStackCloudDefaultStorageAccountType, nil
+		}
+		return azurePublicCloudDefaultStorageAccountType, nil
+	}
+
+	sku := compute.DiskStorageAccountTypes(storageAccountType)
+	supportedSkuNames := compute.PossibleDiskStorageAccountTypesValues()
+	if IsAzureStackCloud(cloud, disableAzureStackCloud) {
+		supportedSkuNames = []compute.DiskStorageAccountTypes{compute.StandardLRS, compute.PremiumLRS}
+	}
+	for _, s := range supportedSkuNames {
+		if sku == s {
+			return sku, nil
+		}
+	}
+
+	return "", fmt.Errorf("azureDisk - %s is not supported sku/storageaccounttype. Supported values are %s", storageAccountType, supportedSkuNames)
+}
+
+// PickAvailabilityZone selects 1 zone given topology requirement.
+// if not found or topology requirement is not zone format, empty string is returned.
+func PickAvailabilityZone(requirement *csi.TopologyRequirement, region, topologyKey string) string {
+	if requirement == nil {
+		return ""
+	}
+	for _, topology := range requirement.GetPreferred() {
+		if zone, exists := topology.GetSegments()[WellKnownTopologyKey]; exists {
+			if IsValidAvailabilityZone(zone, region) {
+				return zone
+			}
+		}
+		if zone, exists := topology.GetSegments()[topologyKey]; exists {
+			if IsValidAvailabilityZone(zone, region) {
+				return zone
+			}
+		}
+	}
+	for _, topology := range requirement.GetRequisite() {
+		if zone, exists := topology.GetSegments()[WellKnownTopologyKey]; exists {
+			if IsValidAvailabilityZone(zone, region) {
+				return zone
+			}
+		}
+		if zone, exists := topology.GetSegments()[topologyKey]; exists {
+			if IsValidAvailabilityZone(zone, region) {
+				return zone
+			}
+		}
+	}
+	return ""
+}
+
+func checkDiskName(diskName string) bool {
+	length := len(diskName)
+
+	for i, v := range diskName {
+		if !(unicode.IsLetter(v) || unicode.IsDigit(v) || v == '_' || v == '.' || v == '-') ||
+			(i == 0 && !(unicode.IsLetter(v) || unicode.IsDigit(v))) ||
+			(i == length-1 && !(unicode.IsLetter(v) || unicode.IsDigit(v) || v == '_')) {
+			return false
+		}
+	}
+
+	return true
+}
+
+func getKubeClient(kubeconfig string) (*kubernetes.Clientset, error) {
+	var (
+		config *rest.Config
+		err    error
+	)
+	if kubeconfig != "" {
+		if config, err = clientcmd.BuildConfigFromFlags("", kubeconfig); err != nil {
+			return nil, err
+		}
+	} else {
+		if config, err = rest.InClusterConfig(); err != nil {
+			return nil, err
+		}
+	}
+
+	return kubernetes.NewForConfig(config)
+}
