+++ test/e2e/dynamic_provisioning_test.go
@@ -21,7 +21,6 @@ import (
 	"fmt"
 	"os"
 
-	"sigs.k8s.io/azuredisk-csi-driver/pkg/azuredisk"
 	"sigs.k8s.io/azuredisk-csi-driver/test/e2e/driver"
 	"sigs.k8s.io/azuredisk-csi-driver/test/e2e/testsuites"
 
@@ -35,6 +34,7 @@ import (
 	clientset "k8s.io/client-go/kubernetes"
 	restclientset "k8s.io/client-go/rest"
 	"k8s.io/kubernetes/test/e2e/framework"
+	constants "sigs.k8s.io/azuredisk-csi-driver/pkg/azureutils"
 )
 
 var _ = ginkgo.Describe("Dynamic Provisioning", func() {
@@ -965,7 +965,7 @@ func (t *dynamicProvisioningTestSuite) normalizeVolume(volume testsuites.VolumeD
 	case "kubernetes.io/azure-disk":
 		volumeBindingMode := storagev1.VolumeBindingWaitForFirstConsumer
 		volume.VolumeBindingMode = &volumeBindingMode
-	case "", azuredisk.DefaultDriverName:
+	case "", constants.DefaultDriverName:
 		if !isMultiZone {
 			return volume
 		}
